Move wallet implements from wallet/core subfolder
