+++ src/wallet/bip44wallet/wallet_test.go
