+++ src/wallet/collection/testdata/scrypt-chacha20poly1305-encrypted.wlt
