+++ src/wallet/collection/testdata/wallet_serialize.wlt
