+++ src/wallet/collection/wallet_test.go
