+++ src/wallet/deterministic/testdata/scrypt-chacha20poly1305-encrypted.wlt
