+++ src/wallet/deterministic/testdata/wallet_serialize.wlt
