+++ src/wallet/deterministic/wallet_test.go
