+++ src/wallet/xpubwallet/wallet_test.go
